importccl: move import reader proc to own file
Release note: none.
